Middle start
We start committing in the middle of this project. Colab notebook almost done. It can have much more comments. we can add them later. main.py is the one runner file of the project. It should alongside with the trained model and a test.png image (For now). In this part of project we are trying to make a UI for our application.